// src/libs/enums/TechEnum.ts
export enum TechEnum {
  // languages
  JAVASCRIPT = 'JAVASCRIPT',
  TYPESCRIPT = 'TYPESCRIPT',
  PYTHON = 'PYTHON',
  JAVA = 'JAVA',
  CSHARP = 'CSHARP',
  POSTGRESQL = 'POSTGRESQL',

  // frontend
  REACT = 'REACT',
  VUE = 'VUE',
  ANGULAR = 'ANGULAR',

  // backend
  NESTJS = 'NESTJS',
  DJANGO = 'DJANGO',
  SPRING_BOOT = 'SPRING_BOOT',
  DOT_NET = 'DOT_NET',

  // DevOps
  DOCKER = 'DOCKER',
  KUBERNETES = 'KUBERNETES',
  JENKINS = 'JENKINS',
  GIT = 'GIT',
}